job-portal-system/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
comtypes==1.4.12
distlib==0.4.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
packaging==25.0
pipenv==2025.0.4
platformdirs==4.3.8
PyJWT==2.11.0
pypiwin32==223
pywin32==311
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.33.0
